Diplom_2 Edits 1
