+++ src/main/java/ru/yandex/praktikum/client/RestClient.java
@@ -1,3 +1,5 @@
+package ru.yandex.praktikum.client;
+
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.http.ContentType;
 import io.restassured.specification.RequestSpecification;
