+++ src/test/java/CreateOrderTest.java
@@ -3,6 +3,9 @@ import io.restassured.response.Response;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import ru.yandex.praktikum.client.AuthClient;
+import ru.yandex.praktikum.client.IngredientsClient;
+import ru.yandex.praktikum.client.OrdersClient;
 import ru.yandex.praktikum.model.Order;
 import ru.yandex.praktikum.model.User;
 import ru.yandex.praktikum.model.UserCredentials;
@@ -18,15 +21,18 @@ public class CreateOrderTest {
     private Response responseAuth;
     private Response responseRegister;
     private Order order;
-    private Client client;
+    private AuthClient authClient;
+    private IngredientsClient ingredientsClient;
+    private OrdersClient ordersClient;
     private User userWithEmailAndPassword = new User();
     private User user;
     private String accessToken;
 
-
     @Before
     public void setUp() {
-        client = new Client();
+        authClient = new AuthClient();
+        ingredientsClient = new IngredientsClient();
+        ordersClient = new OrdersClient();
         order = new Order();
     }
 
@@ -34,36 +40,35 @@ public class CreateOrderTest {
     public void tearDown() {
         if (accessToken != null) {
             accessToken = responseRegister.path("accessToken");
-            client.deleteUser(accessToken);
+            authClient.deleteUser(accessToken);
         }
     }
 
     @Test
     @DisplayName("Создание заказа неавторизованным пользователем c ингредиентами")
     public void createOrderUserWithoutAuthTest() {
-        responseGetIngredient = client.getIngredients();
+        responseGetIngredient = ingredientsClient.getIngredients();
         order.setIngredients(order.listIngredient(responseGetIngredient));
-        responseCreateOrder = client.createOrder(order);
-        assertEquals("Error create order", true, responseCreateOrder.path("success"));
+        responseCreateOrder = ordersClient.createOrder(order);
         assertEquals("Inccorect status code", 200, responseCreateOrder.statusCode());
+        assertEquals("Error create order", true, responseCreateOrder.path("success"));
 
     }
-
 
     @Test
     @DisplayName("Создание заказа авторизованным пользователем c ингредиентами")
     public void createOrderUserAuthTest() {
         user = User.getRandom();
-        responseRegister = client.register(user);
+        responseRegister = authClient.register(user);
         userWithEmailAndPassword.setEmail(user.email);
         userWithEmailAndPassword.setPassword(user.password);
-        responseAuth = client.auth(UserCredentials.from(userWithEmailAndPassword));
+        responseAuth = authClient.auth(UserCredentials.from(userWithEmailAndPassword));
         accessToken = responseAuth.path("accessToken");
-        responseGetIngredient = client.getIngredientsWithAuthToken(accessToken);
+        responseGetIngredient = ingredientsClient.getIngredientsWithAuthToken(accessToken);
         order.setIngredients(order.listIngredient(responseGetIngredient));
-        responseCreateOrder = client.createOrderAuthUser(order, accessToken);
-        assertEquals("Error create order", true, responseCreateOrder.path("success"));
+        responseCreateOrder = ordersClient.createOrderAuthUser(order, accessToken);
         assertEquals("Inccorect status code", 200, responseCreateOrder.statusCode());
+        assertEquals("Error create order", true, responseCreateOrder.path("success"));
 
     }
 
@@ -71,16 +76,16 @@ public class CreateOrderTest {
     @DisplayName("Создание заказа авторизованным пользователем без ингредиентов")
     public void createOrderUserAuthWithoutIngredientsTest() {
         user = User.getRandom();
-        responseRegister = client.register(user);
+        responseRegister = authClient.register(user);
         userWithEmailAndPassword.setEmail(user.email);
         userWithEmailAndPassword.setPassword(user.password);
-        responseAuth = client.auth(UserCredentials.from(userWithEmailAndPassword));
+        responseAuth = authClient.auth(UserCredentials.from(userWithEmailAndPassword));
         accessToken = responseAuth.path("accessToken");
         ArrayList<String> listIngredient = new ArrayList<>();
         order.setIngredients(listIngredient);
-        responseCreateOrder = client.createOrderAuthUser(order, accessToken);
-        assertEquals("Error create order", false, responseCreateOrder.path("success"));
+        responseCreateOrder = ordersClient.createOrderAuthUser(order, accessToken);
         assertEquals("Inccorect status code", 400, responseCreateOrder.statusCode());
+        assertEquals("Error create order", false, responseCreateOrder.path("success"));
 
     }
 
@@ -88,15 +93,15 @@ public class CreateOrderTest {
     @DisplayName("Создание заказа авторизованным пользователем с неверным хеш ингридиентов")
     public void createOrderUserAuthWithIncorrectHashIngredientsTest() {
         user = User.getRandom();
-        responseRegister = client.register(user);
+        responseRegister = authClient.register(user);
         userWithEmailAndPassword.setEmail(user.email);
         userWithEmailAndPassword.setPassword(user.password);
-        responseAuth = client.auth(UserCredentials.from(userWithEmailAndPassword));
+        responseAuth = authClient.auth(UserCredentials.from(userWithEmailAndPassword));
         accessToken = responseAuth.path("accessToken");
         ArrayList<String> listIngredient = new ArrayList<>();
         listIngredient.add("12345");
         order.setIngredients(listIngredient);
-        responseCreateOrder = client.createOrderAuthUser(order, accessToken);
+        responseCreateOrder = ordersClient.createOrderAuthUser(order, accessToken);
         assertEquals("Inccorect status code", 500, responseCreateOrder.statusCode());
 
     }
