+++ src/test/java/UserAuthTest.java
@@ -5,26 +5,25 @@ import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import ru.yandex.praktikum.client.AuthClient;
 import ru.yandex.praktikum.model.User;
 import ru.yandex.praktikum.model.UserCredentials;
 
 import static org.junit.Assert.assertEquals;
 
-
 public class UserAuthTest {
 
     private Response responseRegister;
     private Response responseAuth;
-    private Client client;
+    private AuthClient client;
     private Faker faker = new Faker();
     private User user;
     private User userWithEmailAndPassword = new User();
     private String accessToken;
 
-
     @Before
     public void setUp() {
-        client = new Client();
+        client = new AuthClient();
         user = User.getRandom();
         responseRegister = client.register(user);
         userWithEmailAndPassword.setEmail(user.email);
@@ -44,8 +43,9 @@ public class UserAuthTest {
         userWithEmailAndPassword.setEmail(user.email);
         userWithEmailAndPassword.setPassword(user.password);
         responseAuth = client.auth(UserCredentials.from(userWithEmailAndPassword));
-        assertEquals("Auth error", true, responseAuth.then().extract().path("success"));
         assertEquals("Status code incorrect", 200, responseAuth.statusCode());
+        assertEquals("Auth error", true, responseAuth.then().extract().path("success"));
+
     }
 
     @Test
@@ -54,8 +54,9 @@ public class UserAuthTest {
         userWithEmailAndPassword.setEmail(RandomStringUtils.randomAlphabetic(10) + "@" + RandomStringUtils.randomAlphabetic(10) + ".ru");
         userWithEmailAndPassword.setPassword(user.password);
         responseAuth = client.auth(UserCredentials.from(userWithEmailAndPassword));
-        assertEquals("Email or password are incorrect", false, responseAuth.then().extract().path("success"));
         assertEquals("Status code incorrect", 401, responseAuth.statusCode());
+        assertEquals("Email or password are incorrect", false, responseAuth.then().extract().path("success"));
+
     }
 
     @Test
@@ -64,8 +65,9 @@ public class UserAuthTest {
         userWithEmailAndPassword.setEmail(user.email);
         userWithEmailAndPassword.setPassword(RandomStringUtils.randomAlphabetic(10));
         responseAuth = client.auth(UserCredentials.from(userWithEmailAndPassword));
-        assertEquals("Email or password are incorrect", false, responseAuth.then().extract().path("success"));
         assertEquals("Status code incorrect", 401, responseAuth.statusCode());
+        assertEquals("Email or password are incorrect", false, responseAuth.then().extract().path("success"));
+
     }
 
 }
