+++ src/test/java/UserRegisterTest.java
@@ -3,6 +3,7 @@ import io.restassured.response.Response;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import ru.yandex.praktikum.client.AuthClient;
 import ru.yandex.praktikum.model.User;
 
 import static org.junit.Assert.assertEquals;
@@ -10,14 +11,13 @@ import static org.junit.Assert.assertEquals;
 public class UserRegisterTest {
 
     private Response responseRegister;
-    private Client client;
+    private AuthClient client;
     private User user;
     private String accessToken;
 
-
     @Before
     public void setUp() {
-        client = new Client();
+        client = new AuthClient();
         user = User.getRandom();
         responseRegister = client.register(user);
 
@@ -34,8 +34,8 @@ public class UserRegisterTest {
     @Test
     @DisplayName("Пользователя можно зарегистрировать")
     public void userRegisterSuccessTest() {
-        assertEquals("User is not create", true, responseRegister.then().extract().path("success"));
         assertEquals("Status code incorrect", 200, responseRegister.statusCode());
+        assertEquals("User is not create", true, responseRegister.then().extract().path("success"));
 
     }
 
@@ -44,8 +44,8 @@ public class UserRegisterTest {
     public void doubleUserRegisterFailTest() {
         User doubleUser = new User(user.email, user.password, user.name);
         responseRegister = client.register(doubleUser);
-        assertEquals("Unexpected error", false, responseRegister.then().extract().path("success"));
         assertEquals("Status code incorrect", 403, responseRegister.statusCode());
+        assertEquals("Unexpected error", false, responseRegister.then().extract().path("success"));
 
     }
 
@@ -56,8 +56,8 @@ public class UserRegisterTest {
         userWithoutEmail.setPassword(user.password);
         userWithoutEmail.setName(user.name);
         responseRegister = client.register(userWithoutEmail);
-        assertEquals("Unexpected error", false, responseRegister.then().extract().path("success"));
         assertEquals("Status code incorrect", 403, responseRegister.statusCode());
+        assertEquals("Unexpected error", false, responseRegister.then().extract().path("success"));
 
     }
